DONOR-15: Moving the data logic of the page to the container
