+++ client/src/containers/DonationList/propTypes.js
@@ -2,6 +2,7 @@ import { PropTypes } from 'prop-types';
 
 const propTypes = {
   donations: PropTypes.array,
+  getDonations: PropTypes.func,
 };
 
 export {
