--- client/src/pages/Home/propTypes.js
@@ -1,8 +1,0 @@
-import { PropTypes } from 'prop-types';
-
-const propTypes = {
-  getDonations: PropTypes.func,
-  donations: PropTypes.array,
-};
-
-export default propTypes;
